pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nekyb"
version = "1.0.0"
description = "Advanced vulnerability detection and auto-fixing security tool for comprehensive code analysis"
authors = [
    {name = "NoVuneb Soft", email = "security@novuneb.dev"}
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
keywords = ["security", "vulnerability", "scanner", "static-analysis", "devsecops", "sast"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "bandit>=1.7.9",
    "semgrep>=1.88.0",
    "safety>=3.2.7",
    "pylint>=3.2.7",
    "typer>=0.12.5",
    "rich>=13.5.2,<13.6",
    "pydantic>=2.9.2",
    "requests>=2.32.3",
    "gitpython>=3.1.43",
    "jinja2>=3.1.4",
    "pyyaml>=6.0.2",
    "toml>=0.10.2",
    "python-dotenv>=1.0.1",
    "sarif-om>=1.0.4",
    "jschema-to-python>=1.2.3",
    "pygments>=2.18.0",
    "cryptography>=43.0.1",
    "aiohttp>=3.10.5",
    "libcst>=1.4.0",
    "radon>=6.0.1",
    "vulture>=2.11",
    "pyflakes>=3.2.0",
    "packaging>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.8.0",
    "mypy>=1.11.2",
    "ruff>=0.6.5",
]

[project.scripts]
novuneb = "novuneb.cli:main"

[project.urls]
Homepage = "https://github.com/nekyb/novuneb"
Repository = "https://github.com/nekyb/novuneb"
Documentation = "https://docs.novuneb.dev"
"Bug Tracker" = "https://github.com/nekyb/novuneb/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=novuneb --cov-report=term-missing"
